Implemented Steam converter
