+++ Celeste_Launcher_Gui/Windows/SteamConverterWindow.xaml
@@ -1,0 +1,32 @@
+﻿<Window x:Class="Celeste_Launcher_Gui.Windows.SteamConverterWindow"
+             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
+             xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
+             xmlns:local="clr-namespace:Celeste_Launcher_Gui.UserControls"
+             xmlns:p="clr-namespace:Celeste_Launcher_Gui.Properties"
+             mc:Ignorable="d" 
+             WindowStyle="none" ResizeMode="NoResize" UseLayoutRounding="True" Background="Transparent" AllowsTransparency="True"
+             Width="849" Height="308">
+    <Grid>
+        <Grid.Background>
+            <ImageBrush ImageSource="pack://application:,,,/Celeste_Launcher_Gui;component/Resources/MediumDialogue-Background.png"/>
+        </Grid.Background>
+        <Border BorderThickness="22" BorderBrush="Transparent" MouseDown="BorderMoved">
+            <Grid>
+                <TextBlock FontFamily="{StaticResource Lithos-Pro-Regular}" FontSize="26" Foreground="White" HorizontalAlignment="Center" VerticalAlignment="Top"
+                           MaxWidth="800"
+                                Text="{x:Static p:Resources.SteamConverterTitle}"
+                           TextWrapping="Wrap" TextAlignment="Center" Margin="0,60,0,0"/>
+
+                <Button Style="{StaticResource LargeButton}" Content="{x:Static p:Resources.SteamConverterConfirmBtn}" HorizontalAlignment="Center" VerticalAlignment="Bottom" Margin="0,0,0,20" Click="ConfirmBtnClick"/>
+            </Grid>
+        </Border>
+        <StackPanel Orientation="Horizontal" VerticalAlignment="Top" HorizontalAlignment="Right" Margin="0,4,30,0">
+            <local:WindowControl
+                    DefaultIcon="pack://application:,,,/Celeste_Launcher_Gui;component/Resources/Icons/WCCloseNormal.png"
+                    HoverIcon="pack://application:,,,/Celeste_Launcher_Gui;component/Resources/Icons/WCCloseHover.png"
+                    Click="CloseWindow"/>
+        </StackPanel>
+    </Grid>
+</Window>
